+++ lib/alf-engine/alf/engine/compiler.rb
@@ -57,6 +57,13 @@ module Alf
         SetAttr.new(apply(expr.operand), expr.ext, expr)
       end
 
+      def on_frame(expr)
+        offset, limit = expr.offset, expr.limit
+        ordering = unsupported(expr.ordering){ expr.full_ordering }
+        op = Sort.new(apply(expr.operand), ordering, expr)
+        op = Take.new(op, offset, limit, expr)
+      end
+
       def on_page(expr)
         index, size = expr.page_index, expr.page_size
         ordering = unsupported(expr.ordering){ expr.full_ordering }